exists fix

+++ folderapi/delete.go
@@ -4,8 +4,10 @@ import "os"
 
 func Delete(path string, withContent bool) error {
 	if withContent {
-		if err := os.RemoveAll(path); err != nil {
-			return err
+		if Exists(path) {
+			if err := os.RemoveAll(path); err != nil {
+				return err
+			}
 		}
 	}
